[audio] moved engine playback funcs to audio context
